package/balena-engine: bump to version 20.10.26
Like docker-engine and docker-cli, balena-engine uses a non-standard
vendor/ directory structure, so we use the same fixes as we introduced
for docker-engine ((see commit: package/docker-engine: bump version to
v23.0.0):

  - remove the problematic vendor/modules.txt
  - create a go.mod

Also remove the vendor/archive/ sub-tree (fixes a build error) because
Go no longer allows having duplicate packages in GOPATH and GOROOT.

vendor/archive/tar cannot be in vendor/ because it exists in the
standard library.  Since we remove the full vendor/archive/, that is
also fixed.

https://github.com/balena-os/balena-engine/releases/tag/v20.10.26

Signed-off-by: Christian Stewart <christian@paral.in>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>