package/wolfssl: disable assembly when not supported
wolfssl contains some assembly code and its configure.ac script
enables the assembly code depending on the CPU architecture. However,
the detection logic is not sufficient and leads to using the assembly
code in situation where it should not.

Here are two examples:

 - As soon as the architecture is mips64/mips64el, it uses assembly
   code, but that assembly code is not mips64r6 compatible.

 - As soon as the architecture is RISC-V, it uses assembly code, but
   that assembly code uses multiplication instructions, without paying
   attention that the "M" extension may not be available in the RISC-V
   CPU instruction set.

In order to avoid this, we introduce a
BR2_PACKAGE_WOLFSSL_ASM_SUPPORTS hidden boolean to decide when to
enable/disable assembly code. We set an initial value with the
supported CPU architectures, with the exclusion of the problematic
cases described above.

It is pretty likely that this variable will need tuning progressively,
as it is difficult to determine which CPU variants exactly are
supported by the assembly code in wolfssl.

Fixes:

 - MIPS64 case: http://autobuild.buildroot.net/results/43e/43ee1a457cbeb986d958c9a70cac0676377661c0/
 - RISC-V case: http://autobuild.buildroot.net/results/f05/f056ebe7749a69994afba39a9d4967e2774c45ea/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit d8dc5315eb712eca0a5cbf793a6714a47ab6e57e)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>