package/pkg-autotools: add FOO_AUTOPOINT flag
Add a new variable to run autopoint during autoreconf.

This variable depends on FOO_AUTORECONF = YES. It is mutually exclusive
to FOO_GETTEXTIZE. If both are set, FOO_AUTOPOINT is ignored to preserve
previous behavior.

This support is being added per a previous discussion [1].

The usage of gettextize needs to be removed cuz mama says so [2].
  ... is not part of the GNU build system, in the sense that it should
  not be invoked automatically, and not be invoked by someone who doesn’t
  assume the responsibilities of a package maintainer. For the latter
  purpose, a separate tool is provided, see Invoking the autopoint Program [3]

Using gettextize has the unintended consequence of updating the package
based on the version of host-gettext vs what was chosen by the upstream
maintainer. As mama said above, we should use autopoint. Do as she says
to avoid further scolding (gettextize shoehorning and package patching).

[1] https://patchwork.ozlabs.org/project/buildroot/patch/20170827110920.15579-1-aleksander@aleksander.es/
[2] https://www.gnu.org/software/gettext/manual/html_node/gettextize-Invocation.html
[3] https://www.gnu.org/software/gettext/manual/html_node/autopoint-Invocation.html

Signed-off-by: Vincent Fazio <vfazio@gmail.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>