package/pkg-autotools: drop support for FOO_GETTEXTIZE
Remove all support for FOO_GETTEXTIZE and alert users to FOO_AUTOPOINT
as the recommended solution in its stead. We can use the standard
check-deprecated-variable for this, because from a user perspective
FOO_AUTOPOINT is pretty-much a drop-in replacement.

The warnings about FOO_GETTEXTIZE_OPTS are no longer relevant, because
they will only make sense if FOO_GETTEXTIZE was already set.

Signed-off-by: Vincent Fazio <vfazio@gmail.com>
[Arnout:
 - use check-deprecated-variable;
 - remove FOO_GETTEXTIZE_OPTS warnings;
 - remove definition of GETTEXTIZE]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>