package/postgresql: add host-pkgconf dependency
host-pkgconf is mandatory to find lz4 and zstd resulting in the
following build failure since commit
9cd2e6e090449f9d2f371d24152580c038892f3d:

configure: error: in `/home/autobuild/autobuild/instance-5/output-1/build/postgresql-15.2':
configure: error: The pkg-config script could not be found or is too old.  Make sure it
is in your PATH or set the PKG_CONFIG environment variable to the full
path to pkg-config.

Alternatively, you may set the environment variables LZ4_CFLAGS
and LZ4_LIBS to avoid the need to call pkg-config.
See the pkg-config man page for more details.

Fixes:
 - http://autobuild.buildroot.org/results/8744277ebe9910635ef8fe290c8ba4eee420b538

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>