package/stress-ng: bump to version V0.15.04
This commit dropped the patch, included upstream in:
https://github.com/ColinIanKing/stress-ng/commit/5d419c790e648c7a2f96f34ed1b93b326f725545
which was included in V0.14.04.

Three patches are also introduced to fix build issues (all
upstream not but not yet in version).

Also, this new version now depends on BR2_TOOLCHAIN_HAS_SYNC_4.

Finally, a new _MAKE_FLAGS is introduced to prevent the package to
modify too much its CFLAGS.

For change log since V0.13.05, see:
- https://github.com/ColinIanKing/stress-ng/blob/V0.15.04/debian/changelog
or commit logs:
- https://github.com/ColinIanKing/stress-ng/commits/V0.15.04

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>