package/libgpg-error: force cross compiling
By default autotools enters cross-compiling mode if --build and --host
differ [0]. This should normally be the case as buildroot sets itself as
vendor in the toolchain triplet.
For libgpg-error the --host flag has to be set to a generic triplet
because this information is used to lookup the syscfg.
When compiling on aarch64 for aarch64 build and host would be the same,
therefore cross-compiling isn't activated, which results in a build failure.

Fixes:
 - http://autobuild.buildroot.net/results/6d4/6d4bc520c12313508d0fadee11d5b88c35a3cba9/

[0]: https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.71/html_node/Hosts-and-Cross_002dCompilation.html

Signed-off-by: Daniel Lang <d.lang@abatec.at>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>