package/mmc-utils: bump to version d4c2910981ff99b983734426dfa99632fb81ac6b
Let's also switch to a _SITE that can be accessed from a web browser and
results to be the official one for mmc-utils.

The license file we use is a header, so it changes every now and then,
even when the licenins terms do not. Switch to using README instead, as
it identifies the license (albeit just by name, b ut that's enough for
us), which is still GPL-2.0 (only).

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>
[yann.morin.1998@free.fr: switch to README]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>