Revert "package/skeleton: drop $(HOST_DIR)/usr compatibility symlink"
In 38b798da73ba (package/skeleton: drop $(HOST_DIR)/usr compatibility
symlink), we considered that $(HOST_DIR)/usr is a legacy symlink, and so
we dropped it. During review, it was pointed out that:

 1. [if] something installs in HOST_DIR/usr it's probably going to
    break somewhere down the line,

 2. DESTDIR is empty for host builds. And PREFIX/usr would be completely
    crazy.

However, we do have a package where DESTDIR is not empty for host
packages. Indeed, for host-systemd, we do have a convoluted sequence
where we configure it for the traditional /usr prefix, rather than
$(HOST_DIR) for all other host packages, and we install it with
DESTDIR=$(HOST_DIR). This is because systemd has to interpret path on
the target, not on the host, and confusion ensues if host systemd is not
configured with the same prefix as the target systemd. See commit
35c11a027c88 (package/systemd: add host variant) for the full, in-depth
explanations.

This reverts commit 38b798da73ba69c9e1d8279805b621f80b98bc64.

This also changes the comment to note that, although legacy, the usr
symlink is still required for some weird host packages.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>