board/pc/linux.config: Add framebuffer and pci support
Virtually all PCs nowadays have a PCI bridge and PCI devices.

All framnebuffer options depend on CONFIG_FB, which default is not 'y',
so also enable that.

The overwhelming majority of PCs nowadays also come with an UEFI
bootloader, that configures a framebuffer usable in early boot, which
the kernel can reuse without needing any hardware-specific driver.

EDID allows retrieving the capabilities of the connected display,
especially the frequencies and so on, to properly drive the display.

Signed-off-by: Adam Duskett <aduskett@gmail.com>
[yann.morin.1998@free.fr:
  - drop cosmetic CONFIG_FRAMEBUFFER_CONSOLE_DEFERRED_TAKEOVER=y
  - rewrite commit log to explain all new options
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>