package/openocd: add the eCos exception to the GPL-2.0+ license
As noted by Yann E. MORIN in [1] only the src/flash/nand/ecc.c file
refers to the eCos license.

The SPDX identifier for that file is:
SPDX-License-Identifier: GPL-2.0-or-later WITH eCos-exception-2.0

The file is part of the openocd executable via libopenocd.la, which being
a static library it is not installed.

Based on this, becomes more natural to associate the eCos exception with
the GPL 2.0+ license, rather than considering it separately as a license.

[1] https://patchwork.ozlabs.org/project/buildroot/patch/20230411203930.2455301-1-dario.binacchi@amarulasolutions.com/
Suggested-by: Yann E. MORIN <yann.morin.1998@free.fr>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 548e30d140066d5464eef1c3be57284a1a3dcf5e)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>