package/python-pip: add host variant
Some packages, like libselinux, use pip to perform the installation step
of their python bindings. Those packages do have to use our host python
when building, so that it matches the target python.

So we need to have a host variant of pip that is usable with our host
python.

Signed-off-by: Adam Duskett <aduskett@gmail.com>
[yann.morin.1998@free.fr: reword commit log]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>